moves mpr121 to binary_sensor
