+++ src/esphome/application.cpp
@@ -1271,7 +1271,7 @@ sensor::APDS9960 *Application::make_apds9960(uint32_t update_interval) {
 #endif
 
 #ifdef USE_MPR121
-sensor::MPR121_Sensor *Application::make_mpr121(uint8_t address) {
+binary_sensor::MPR121_Sensor *Application::make_mpr121(uint8_t address) {
   return this->register_component(new MPR121_Sensor(this->i2c_, address));
 }
 #endif
