+++ src/esphome/application.h
@@ -37,6 +37,7 @@
 #include "esphome/binary_sensor/rdm6300.h"
 #include "esphome/binary_sensor/status_binary_sensor.h"
 #include "esphome/binary_sensor/template_binary_sensor.h"
+#include "esphome/binary_sensor/mpr121_sensor.h"
 #include "esphome/cover/cover.h"
 #include "esphome/cover/mqtt_cover_component.h"
 #include "esphome/cover/template_cover.h"
@@ -83,7 +84,6 @@
 #include "esphome/sensor/adc.h"
 #include "esphome/sensor/ads1115_component.h"
 #include "esphome/sensor/apds9960.h"
-#include "esphome/sensor/mpr121_sensor.h"
 #include "esphome/sensor/bh1750_sensor.h"
 #include "esphome/sensor/bme280_component.h"
 #include "esphome/sensor/bme680_component.h"
@@ -434,6 +434,10 @@ class Application {
 
 #ifdef USE_CUSTOM_BINARY_SENSOR
   binary_sensor::CustomBinarySensorConstructor *make_custom_binary_sensor(const std::function<std::vector<binary_sensor::BinarySensor *>()> &init);
+#endif
+
+#ifdef USE_MPR121
+  binary_sensor::MPR121_Sensor *make_mpr121(uint8_t address = 0x5A);
 #endif
 
 
@@ -840,10 +844,6 @@ class Application {
 
 #ifdef USE_APDS9960
   sensor::APDS9960 *make_apds9960(uint32_t update_interval = 60000);
-#endif
-
-#ifdef USE_MPR121
-  sensor::MPR121_Sensor *make_mpr121(uint8_t address = 0x5A);
 #endif
 
 #ifdef USE_CUSTOM_SENSOR
